+++ docs/search/variables_1.js
@@ -1,5 +1,5 @@
 var searchData=
 [
   ['coulomb_5fcutoff',['Coulomb_cutoff',['../class_excimontec_1_1_parameters.html#aa58e148db895ecda0e5c144fce268498',1,'Excimontec::Parameters']]],
-  ['current_5fversion',['Current_version',['../namespace_k_m_c___lattice.html#a7da01aab47adc940e19891e5979e3b72',1,'KMC_Lattice']]]
+  ['current_5fversion',['Current_version',['../namespace_k_m_c___lattice.html#a91f4a916d9235dfa664004930666c9fa',1,'KMC_Lattice']]]
 ];